fix: remove utilities folder refactor
