+++ packages/fast-form-generator-react/src/class-name-contracts/form-item.array.class-name-contract.ts
@@ -7,8 +7,8 @@ export default interface FormItemArrayClassNameContract {
     formItemArray_controlAddButton: string;
     formItemArray_controlLabel: string;
     formItemArray_existingItemList: string;
-    formItemArray_existingItemList__sorting: string;
     formItemArray_existingItemListItem: string;
+    formItemArray_existingItemListItem__sorting: string;
     formItemArray_existingItemListItemLink: string;
     formItemArray_existingItemRemoveButton: string;
 }
