+++ packages/fast-form-generator-react/src/form/form-item.array.spec.tsx
@@ -29,9 +29,9 @@ const managedClasses: FormItemArrayClassNameContract = {
     formItemArray_controlAddButton: "formItemArray_controlAddButton-class",
     formItemArray_controlLabel: "formItemArray_controlLabel-class",
     formItemArray_existingItemList: "formItemArray_existingItemList-class",
-    formItemArray_existingItemList__sorting:
-        "formItemArray_existingItemList__sorting-class",
     formItemArray_existingItemListItem: "formItemArray_existingItemListItem-class",
+    formItemArray_existingItemListItem__sorting:
+        "formItemArray_existingItemListItem__sorting-class",
     formItemArray_existingItemListItemLink:
         "formItemArray_existingItemListItemLink-class",
     formItemArray_existingItemRemoveButton:
