+++ packages/fast-form-generator-react/src/form/form-item.textarea.style.ts
@@ -5,7 +5,7 @@ import {
     applyLabelStyle,
     applySoftRemove,
     applySoftRemoveInput,
-} from "../utilities";
+} from "./form.utilities.style";
 import { ComponentStyles } from "@microsoft/fast-jss-manager";
 import { FormItemTextareaClassNameContract } from "../class-name-contracts/";
