+++ packages/fast-form-generator-react/src/index.ts
@@ -2,5 +2,3 @@ import Form from "./form";
 
 export default Form;
 export * from "./form";
-
-export * from "./utilities";
