--- packages/fast-form-generator-react/src/utilities/index.ts
@@ -1,2 +1,0 @@
-export * from "./form.constants.style";
-export * from "./form.utilities.style";
